+++ context-checkers.d/40cmake-build-dir.sh
@@ -36,13 +36,18 @@ function _show_cmake_options()
                 "${_sco__top_build_dir}"/CMakeCache.txt \
               )
             local _sco__color_build_type
-            _get_color_param SP_CMAKE_BUILD_TYPE_COLOR sp_color_notice _sco__color_build_type
+            local _sco__build_type_color_var="SP_CMAKE_${_sco__build_type^^}_BUILD_TYPE_COLOR"
+            if [[ -n ${!_sco__build_type_color_var} ]]; then
+                _get_color_param ${_sco__build_type_color_var} sp_color_notice _sco__color_build_type
+            else
+                _get_color_param SP_CMAKE_BUILD_TYPE_COLOR sp_color_notice _sco__color_build_type
+            fi
             local _sco__color_version
             _get_color_param SP_CMAKE_PROJECT_VERSION_COLOR sp_color_info _sco__color_version
             local _sco__color_install_pfx
             _get_color_param SP_CMAKE_INSTALL_PREFIX_COLOR sp_color_debug _sco__color_install_pfx
 
-            printf "${_sco__build_type:-"${_sco__color_build_type}default"}${sp_seg}${_sco__color_version}${_sco__version}→${_sco__color_install_pfx}${_sco__prefix}"
+            printf "${_sco__color_build_type}${_sco__build_type:-default}${sp_seg}${_sco__color_version}${_sco__version}→${_sco__color_install_pfx}${_sco__prefix}"
         else
             local _sco__color_broken_mark
             _get_color_param SP_CMAKE_BROKEN_MARK_COLOR sp_color_alert _sco__color_broken_mark