+++ ChangeLog.md
@@ -12,7 +12,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Add options `USE_GIT` and `USE_SVN` with default value `AUTO` to enforce
   install corresponding checker, even if no program has found;
 - Add option `USE_GENTOO` with default value `AUTO` to install Gentoo
-  specific checkers.
+  specific checkers;
+- Colorize CMake build types. One can set `SP_CMAKE_<UPCASE_NAME>_BUILD_TYPE_COLOR`
+  to desired color. If not set, default `SP_CMAKE_BUILD_TYPE_COLOR` will be used.
 
 ### Changed